pyproject.toml
[project]
name = "ag-cli"
version = "0.1.0"
description = "Antigravity-style agent/orchestrator CLI (Python)"
requires-python = ">=3.11"
dependencies = [
  "typer>=0.12.0",
  "rich>=13.7.0",
  "pydantic>=2.8.0",
  "pyyaml>=6.0.1",
  "python-dotenv>=1.0.1",
  "openai>=1.40.0",
  "playwright>=1.46.0"
]

[project.scripts]
ag = "agcli.cli:app"

[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["agcli"]
